se corrigio el manejo de excepciones mediante @RestControllerAdvice (Cliente,Producto,Venta,Detalle), se cambio la respuesta en los endpoints por ResponseEntitys teniendo un gran manejo en las respuestas Http. Se cambiaron las rutas de los endpoints para tener una mayor compresion de las mismas y se agrego la clase ExceptionUtils

+++ proyectoventasbazar/src/main/java/com/proyecto/ventas/proyectoventasbazar/service/ClienteService.java
@@ -1,8 +1,11 @@
 package com.proyecto.ventas.proyectoventasbazar.service;
 
+import com.proyecto.ventas.proyectoventasbazar.exceptions.ExceptionUtils;
+import com.proyecto.ventas.proyectoventasbazar.exceptions.InvalidArgumentException;
 import com.proyecto.ventas.proyectoventasbazar.exceptions.ResourceNotFoundException;
 import com.proyecto.ventas.proyectoventasbazar.model.Cliente;
 import com.proyecto.ventas.proyectoventasbazar.repository.IClienteRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,33 +20,41 @@ public class ClienteService implements IClienteService {
     IClienteRepository clienteRepo;
 
     @Override
-    public List<Cliente> getClientes() {
-        return clienteRepo.findAll();
+    public List<Cliente> getClientes() throws ResourceNotFoundException {
+        List<Cliente> clientes = clienteRepo.findAll();
+        ExceptionUtils.validateListNotEmpty(clientes, "No se han cargado clientes al sistema");
+        
+        return clientes;
     }
 
     @Override
     public Cliente findCliente(Long idCliente) throws ResourceNotFoundException {
         Optional<Cliente> cliente = clienteRepo.findById(idCliente);
         if(!cliente.isPresent()){
-            throw new ResourceNotFoundException("Cliente no encontrado", "P-404");
+            throw new ResourceNotFoundException("Cliente con la id " + idCliente + " no encontrado" , "P-404");
         }
         return cliente.get();
               
     }
     
-    
 
     @Override
+    @Transactional
     public void saveCliente(Cliente cliente) {
         clienteRepo.save(cliente);
     }
 
     @Override
-    public void deleteCliente(Long idCliente) {
+    @Transactional
+    public void deleteCliente(Long idCliente) throws ResourceNotFoundException {
+        if(!clienteRepo.existsById(idCliente)){
+              throw new ResourceNotFoundException("Cliente con la id " + idCliente +  " no encontrado", "P-404");
+        }
         clienteRepo.deleteById(idCliente);
     }
 
     @Override
+    @Transactional
     public void editCliente(Long idCliente, Cliente cliente) {
         Cliente clienteEditar = this.findCliente(idCliente);
         clienteEditar.setNombre(cliente.getNombre());
@@ -52,5 +63,6 @@ public class ClienteService implements IClienteService {
         clienteEditar.setDni(cliente.getDni());
         clienteRepo.save(clienteEditar);
     }
-   
+    
+    
 }
