+++ proyectoventasbazar/src/main/java/com/proyecto/ventas/proyectoventasbazar/service/ProductoService.java
@@ -1,9 +1,13 @@
 package com.proyecto.ventas.proyectoventasbazar.service;
 
 import com.proyecto.ventas.proyectoventasbazar.dto.DetalleDTO;
+import com.proyecto.ventas.proyectoventasbazar.exceptions.EmptyListException;
+import com.proyecto.ventas.proyectoventasbazar.exceptions.ExceptionUtils;
+import com.proyecto.ventas.proyectoventasbazar.exceptions.ResourceNotFoundException;
 import com.proyecto.ventas.proyectoventasbazar.model.DetalleVenta;
 import com.proyecto.ventas.proyectoventasbazar.model.Producto;
 import com.proyecto.ventas.proyectoventasbazar.repository.IProductoRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,26 +21,35 @@ public class ProductoService implements IProductoService {
     IProductoRepository producRepo;
 
     @Override
-    public List<Producto> getProductos() {
-        return producRepo.findAll();
+    public List<Producto> getProductos() throws ResourceNotFoundException  {
+        List<Producto> productos = producRepo.findAll();
+        ExceptionUtils.validateListNotEmpty(productos, "No se han cargado productos en el sistema");
+        return productos;
     }
 
     @Override
-    public Producto findProducto(Long codigoProducto) {
-        return producRepo.findById(codigoProducto).orElseThrow(()->new RuntimeException("Producto no encontrado"));
+    public Producto findProducto(Long codigoProducto) throws ResourceNotFoundException {
+        return producRepo.findById(codigoProducto).orElseThrow(()->new ResourceNotFoundException("Producto con codigo" + codigoProducto + " no encontrado"
+                                   ,"P-404"));
     }
 
     @Override
+    @Transactional
     public void saveProducto(Producto producto) {
         producRepo.save(producto);
     }
 
     @Override
-    public void deleteProducto(Long codigoProducto) {
+    @Transactional
+    public void deleteProducto(Long codigoProducto) throws ResourceNotFoundException {
+        if(!producRepo.existsById(codigoProducto)){
+            throw new ResourceNotFoundException("Producto con codigo " + codigoProducto + " no encontrado", "P-404" );
+        }
         producRepo.deleteById(codigoProducto);
     }
 
     @Override
+    @Transactional
     public void editProducto(Long codigoProducto,Producto producto) {
         Producto productoEditar = this.findProducto(codigoProducto);
         productoEditar.setNombre(producto.getNombre());
@@ -48,20 +61,20 @@ public class ProductoService implements IProductoService {
 
 
     @Override
-    public List<Producto> getStockMen5(){
+    public List<Producto> getStockMenorA(int stock) throws EmptyListException{
 
-        List<Producto> prodStockMen5 = new ArrayList<>();
+        List<Producto> prodStockMenor = new ArrayList<>();
         for(Producto prod : producRepo.findAll()){
-            if(prod.getStock()<5){
-                prodStockMen5.add(prod);
+            if(prod.getStock()< stock){
+                prodStockMenor.add(prod);
             }
         }
-
-        return prodStockMen5;
+        ExceptionUtils.validateListNotEmpty(prodStockMenor, "Productos con stock menor a" + stock +  "no encontrados");
+        return prodStockMenor;
     }
 
     @Override
-    public List<DetalleDTO> getDetallesporProducto(Long codigoProducto) {
+    public List<DetalleDTO> getDetallesporProducto(Long codigoProducto) throws ResourceNotFoundException {
         Producto producto = this.findProducto(codigoProducto);
         List<DetalleDTO> detalles = new ArrayList<>();
         for (DetalleVenta detalle : producto.getDetallesProduc()){
@@ -70,9 +83,12 @@ public class ProductoService implements IProductoService {
             detalles.add(detalleDTO);
 
         }
-
+        ExceptionUtils.validateListNotEmpty(detalles, "no se han cargado detalles del producto con codigo" + codigoProducto);
         return detalles;
     }
+    
+    
+    
 
 
 }
